pyproject.toml
[project]
name = "ladon-hackernews"
version = "0.0.1"
description = "Hacker News adapter for the Ladon crawler framework."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}

dependencies = [
    # ladon-crawl is the PyPI name for the Ladon core framework.
    # Until it is published, install from source:
    #   pip install git+https://github.com/MoonyFringers/ladon.git
    "ladon-crawl>=0.0.1",
    "duckdb>=1.0.0",
    "pytz>=2023.3",  # required by DuckDB for TIMESTAMPTZ handling
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "pre-commit",
]

[project.scripts]
ladon-hackernews = "ladon_hackernews.cli:main"

[project.urls]
Repository = "https://github.com/MoonyFringers/ladon-hackernews"
Framework  = "https://github.com/MoonyFringers/ladon"
Issues     = "https://github.com/MoonyFringers/ladon-hackernews/issues"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

# ----------------------------------------
# BLACK
# ----------------------------------------
[tool.black]
line-length = 80
target-version = ['py312', 'py313']

# ----------------------------------------
# ISORT
# ----------------------------------------
[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 80

# ----------------------------------------
# RUFF
# ----------------------------------------
[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = ["E", "F", "W", "B"]
ignore = [
    "E501",  # Handled by Black
]

# ----------------------------------------
# PYRIGHT
# ----------------------------------------
[tool.pyright]
typeCheckingMode = "strict"
include = ["src", "tests"]
extraPaths = ["src"]
reportMissingImports = "none"
reportUntypedFunctionDecorator = "none"

# ----------------------------------------
# PYTEST
# ----------------------------------------
[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra --tb=short"

# ----------------------------------------
# COVERAGE
# ----------------------------------------
[tool.coverage.run]
source = ["ladon_hackernews"]
branch = true
omit = ["*/ladon_hackernews/__main__.py"]

[tool.coverage.report]
fail_under = 75
show_missing = true
